basic login should work
login works with the sqlhandler and the user.java to create a new user.

Some things are uncertain such as what format the date for dob will save as - but once erin has done the registering i can adjust the login as needed and validate.

added sqlite jar for database connection which changed project properties. should launch for everyone