had to account for us being in BST
when the user would last go to the QnA place, it would store their current time in BST converted to milliseconds to store in the database.
The question and replies time were stored in UTC which is BST - 1.

This resulted in the user last visiting the QNA 1 hour into the future which would result in the SQL query returning 0 because it would think since its in the database as 1 hour in the future then the user has already seen the latest reply